Gallery Page OverHaul
This includes a new Control bar at the bottom of the screen that replaces the previous view button. It also has a back and forth button, and the overlay to view all the images also incorporate this controlbar. It look just like the navbar, same styling and all. This makes the rest of the website consistent, and going away with the old buttons makes the site feel professional.